requirements.txt
python >= 3.9.0
streamlit==1.21.0
opencv-python==4.9.0.80
mediapipe==0.10.11
numpy==1.23.2
mpmath==1.2.1
boto3==1.35.28
praat-parselmouth==0.4.5
audio-extract==0.7.0
requests
